requirements.txt
langchain
langchain-community
langchain-huggingface
faiss-cpu
pypdf
huggingface_hub
flask
python-dotenv
